+++ scripts/dumpRobloxTypes.py
@@ -82,8 +82,8 @@ LUAU_SNIPPET_PATCHES = {
 }
 
 TYPE_INDEX = {
-    "Tuple": "any",
     "Variant": "any",
+    "Tuple": "...any",
     "Function": "((...any) -> ...any)",
     "function": "((...any) -> ...any)",
     "bool": "boolean",
@@ -157,6 +157,12 @@ EXTRA_MEMBERS = {
         "function __unm(self): UDim",
     ],
     "CFrame": [
+        "function ToWorldSpace(self, ...: CFrame): ...CFrame",
+        "function ToObjectSpace(self, ...: CFrame): ...CFrame",
+        "function PointToWorldSpace(self, ...: Vector3): ...Vector3",
+        "function PointToObjectSpace(self, ...: Vector3): ...Vector3",
+        "function VectorToWorldSpace(self, ...: Vector3): ...Vector3",
+        "function VectorToObjectSpace(self, ...: Vector3): ...Vector3",
         "function __add(self, other: Vector3): CFrame",
         "function __sub(self, other: Vector3): CFrame",
         "function __mul(self, other: CFrame): CFrame",
@@ -1017,20 +1023,42 @@ def declareClass(klass: Union[ApiClass, DataType]) -> str:
 
     def declareMember(member: ApiMember):
         if member["MemberType"] == "Property":
-            return (
-                f"\t{escapeName(member['Name'])}: {resolveType(member['ValueType'])}\n"
-            )
-        elif member["MemberType"] == "Function":
-            return f"\t{resolveDeprecation(member, klass)}function {escapeName(member['Name'])}(self{', ' if len(member['Parameters']) > 0 else ''}{resolveParameterList(member['Parameters'])}): {resolveReturnType(member)}\n"
-        elif member["MemberType"] == "Event":
-            parameters = ", ".join(
-                map(lambda x: resolveType(x["Type"]), member["Parameters"])
-            )
-            return f"\t{escapeName(member['Name'])}: RBXScriptSignal<{parameters}>\n"
-        elif member["MemberType"] == "Callback":
-            return f"\t{escapeName(member['Name'])}: ({resolveParameterList(member['Parameters'])}) -> {resolveReturnType(member)}\n"
+            return f"\t{escapeName(member['Name'])}: {resolveType(member['ValueType'])}\n"
         else:
-            assert False, "Unhandled member type: " + member["MemberType"]
+            # only one pack allowed, normalize all but the last to 'any'
+            types = [resolveType(param["Type"]) for param in member["Parameters"]]
+            packIndices = [i for i, t in enumerate(types) if t.startswith("...")]
+
+            for i in packIndices[:-1]:  
+                types[i] = "any"
+
+            # pack must be the last parameter, otherwise normalize it to 'any'.
+            packIndex = next((i for i, t in enumerate(types) if t.startswith("...")), -1)
+
+            if packIndex >= 0 and packIndex != len(types) - 1:
+                types[packIndex] = "any"
+
+            if member["MemberType"] == "Event":
+                typeList = ", ".join(types)
+                typeList = f"({typeList})" if len(types) != 1 else typeList
+                return f"\t{escapeName(member['Name'])}: RBXScriptSignal<{typeList}>\n"
+            else:
+                types = [param["Type"] for param in member["Parameters"]]
+                tuples = [ty for i, ty in enumerate(types) if types[i] == "any"]
+                
+                # Normalize 'any' types to their C++ 'Variant' type.
+                for i, tuple in enumerate(tuples):
+                    tuple["Category"] = "Group"
+                    tuple["Name"] = "Variant"
+                
+                params = resolveParameterList(member["Parameters"])
+                
+                if member["MemberType"] == "Function":
+                    return f"\t{resolveDeprecation(member, klass)}function {escapeName(member['Name'])}(self{', ' if len(member['Parameters']) > 0 else ''}{params}): {resolveReturnType(member)}\n"
+                elif member["MemberType"] == "Callback":
+                    return f"\t{escapeName(member['Name'])}: ({params}) -> {resolveReturnType(member)}\n"
+                else:
+                    assert False, "Unhandled member type: " + member["MemberType"]
 
     memberDefinitions = [
         declareMember(m) for m in klass["Members"] if filterMember(klass["Name"], m)